+++ libs/Renderer/Driver/OpenGL/Shaders/header/Lights.glsl
@@ -67,7 +67,7 @@ INLINE float PointLightIntensity(
     IN(float) a_Falloff)
 {
     float distAtt  = 1.f / pow(a_Distance, a_Falloff);
-    float rangeAtt = isinf(a_Range) ? 1 : max(min(1.f - pow(a_Distance / a_Range, 4.f), 1.f), 0.f) / a_Distance * a_Distance;
+    float rangeAtt = isinf(a_Range) ? 1 : max(min(1.f - float(pow(a_Distance / a_Range, 4.f)), 1.f), 0.f) / a_Distance * a_Distance;
     return a_MaxIntensity * distAtt * rangeAtt;
 }
